fix(ci): trim stale inbox wrappers
